topfish rescale_layout: push down 'size_t' changes
Allows removing some cast compromises from an earlier commit.

+++ lib/topfish/hierarchy.h
@@ -86,7 +86,7 @@ int init_ex_graph(v_data * graph1, v_data * graph2, int n,
 
 // layout distortion:
 void rescale_layout(double *x_coords, double *y_coords,
-    int n, int interval, double width, double height,
+    size_t n, int interval, double width, double height,
     double margin, double distortion);
 void rescale_layout_polar(double * x_coords, double * y_coords, 
     double * x_foci, double * y_foci, int num_foci, size_t n, int interval,
